Fix audio timestamp not start with 0

+++ libfqrtmp/jni/common.cpp
@@ -4,7 +4,7 @@
 #define THREAD_NAME "common"
 extern JNIEnv *jni_get_env(const char *name);
 
-Packet::Packet(uint8_t *data_, int size_, int pts_, int dts_) :
+Packet::Packet(uint8_t *data_, int size_, uint64_t pts_, uint64_t dts_) :
     data(NULL), size(size_), pts(pts_), dts(dts_)
 {
     if (size > 0) {
