+++ libfqrtmp/jni/common.h
@@ -25,9 +25,9 @@ struct Rational {
 struct Packet {
     uint8_t *data;
     int size;
-    int64_t pts, dts;
+    uint64_t pts, dts;
 
-    explicit Packet(uint8_t *data_ = NULL, int size_ = 0, int pts_ = 0, int dts_ = 0);
+    explicit Packet(uint8_t *data_ = NULL, int size_ = 0, uint64_t pts_ = 0, uint64_t dts_ = 0);
     ~Packet();
 };
